Video Handler re-written to use a simpler, more structure, more OO method. The handler maintains an array list of video objects, and each video object maintains its own state and controls etc. #45
